fix(core): MySQL last_sync_at as ISO text; 64-bit version columns; Date-safe lastSyncAt
MySQL DATETIME rejects the trailing Z of the ISO string SqlStore writes
(strict mode), so a MySQL site failed its first sync. The column is now
VARCHAR(40) like SQLite's TEXT. version becomes BIGINT/bigint so any
future epoch-style version cannot overflow, and lastSyncAt() normalises
the Date a Postgres driver returns for timestamptz.

Found by the Nishany integration (task C3).

+++ packages/core-js/src/migrations/postgres.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS seo_runtime_state (
   id integer PRIMARY KEY CHECK (id = 1),
-  version integer NOT NULL DEFAULT 0,
+  version bigint NOT NULL DEFAULT 0,
   site_slug text NOT NULL DEFAULT '',
   settings jsonb NOT NULL DEFAULT '{}',
   last_sync_at timestamptz
